requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-multipart
requests